Rename file to match naming conventions.
